Publication comment tests (#21)
* optimize-services&queries
add author checks in related methods

* Update Exception Messages

* return http responses updated

* Added test setup, dbMock, and tests

* Added tests for each service method

* Fix interface, added tests

* optimize-services&queries
add author checks in related methods

* Update Exception Messages

* return http responses updated

* Added test setup, dbMock, and tests

* Added tests for each service method

* Fix interface, added tests

* Change userId

* fix-work with new method

+++ AI-Social-Platform.Services.Data/PublicationService.cs
@@ -2,32 +2,36 @@
 using AI_Social_Platform.Data.Models.Publication;
 using AI_Social_Platform.Services.Data.Interfaces;
 using AI_Social_Platform.Services.Data.Models.PublicationDtos;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
+
+using static AI_Social_Platform.Common.ExceptionMessages.PublicationExceptionMessages;
 
 namespace AI_Social_Platform.Services.Data;
 
 public class PublicationService : IPublicationService
 {
     private readonly ASPDbContext dbContext;
+    private readonly HttpContext httpContext;
 
-    public PublicationService(ASPDbContext dbContext)
+    public PublicationService(ASPDbContext dbContext, IHttpContextAccessor accessor)
     {
         this.dbContext = dbContext;
+        httpContext = accessor.HttpContext!;
     }
     public async Task<IEnumerable<PublicationDto>> GetPublicationsAsync()
     {
         var publications = await
             dbContext.Publications
-                .OrderByDescending(p => p.DateCreated)
                 .Select(p => new PublicationDto()
                 {
                     Id = p.Id,
                     Content = p.Content,
-                    DateCreated = p.DateCreated.ToShortDateString() 
-                                  + " " 
-                                  + p.DateCreated.ToShortTimeString(),
+                    DateCreated = p.DateCreated,
                     AuthorId = p.AuthorId
                 })
+                .OrderByDescending(p => p.DateCreated)
                 .ToListAsync();
 
         return publications;
@@ -36,21 +40,18 @@ public class PublicationService : IPublicationService
     public async Task<PublicationDto> GetPublicationAsync(Guid id)
     {
         var publication = await dbContext.Publications
-            .Where(p => p.Id == id)
-            .FirstOrDefaultAsync();
+            .FirstOrDefaultAsync(p => p.Id == id);
 
         if (publication == null)
         {
-            throw new Exception("Publication not found");
+            throw new NullReferenceException(PublicationNotFound);
         }
 
         PublicationDto publicationDto = new()
         {
             Id = publication.Id,
             Content = publication.Content,
-            DateCreated = publication.DateCreated.ToShortDateString()
-                          + " " 
-                          + publication.DateCreated.ToShortTimeString(),
+            DateCreated = publication.DateCreated,
             AuthorId = publication.AuthorId
         };
 
@@ -60,11 +61,12 @@ public class PublicationService : IPublicationService
     
     public async Task CreatePublicationAsync(PublicationFormDto dto)
     {
-       Publication publication = new()
-       {
+        var userId = await GetUserId();
+        Publication publication = new()
+        {
            Content = dto.Content,
-           AuthorId = dto.AuthorId
-       };
+           AuthorId = userId
+        };
 
        await dbContext.AddAsync(publication);
        await dbContext.SaveChangesAsync();
@@ -73,15 +75,16 @@ public class PublicationService : IPublicationService
     public async Task UpdatePublicationAsync(PublicationFormDto dto, Guid id)
     {
         var publication = await dbContext.Publications.FirstOrDefaultAsync(p => p.Id == id);
+        var userId = await GetUserId();
 
         if (publication == null)
         {
-            throw new Exception("Publication not found");
+            throw new NullReferenceException(PublicationNotFound);
         }
 
-        if (publication.AuthorId != dto.AuthorId)
+        if (publication.AuthorId != userId)
         {
-            throw new Exception("You are not the author of this publication");
+            throw new AccessViolationException(NotAuthorizedToEditPublication);
         }
 
         publication.Content = dto.Content;
@@ -91,15 +94,20 @@ public class PublicationService : IPublicationService
     public async Task DeletePublicationAsync(Guid id)
     {
         var publication = await dbContext.Publications
-            .Include(p => p.Comments)
             .FirstOrDefaultAsync(p => p.Id == id);
+
+        var userId = await GetUserId();
 
         if (publication == null)
         {
-            throw new Exception("Publication not found");
+            throw new NullReferenceException(PublicationNotFound);
         }
 
-        dbContext.Comments.RemoveRange(publication.Comments);
+        if (publication.AuthorId != userId)
+        {
+            throw new AccessViolationException(NotAuthorizedToDeletePublication);
+        }
+
         dbContext.Publications.Remove(publication);
         await dbContext.SaveChangesAsync();
     }
@@ -109,15 +117,16 @@ public class PublicationService : IPublicationService
     {
         var publication = await dbContext.Publications
             .FirstOrDefaultAsync(p => p.Id == publicationId);
+        var userId = await GetUserId();
 
         if (publication == null)
         {
-            throw new Exception("Publication not found");
+            throw new NullReferenceException(PublicationNotFound);
         }
 
         await dbContext.Comments.AddAsync(new Comment()
         {
-            AuthorId = dto.AuthorId,
+            AuthorId = userId,
             Content = dto.Content,
             PublicationId = publicationId
         });
@@ -127,15 +136,16 @@ public class PublicationService : IPublicationService
     public async Task UpdateCommentAsync(CommentFormDto dto, Guid id)
     {
        var comment = await dbContext.Comments.FirstOrDefaultAsync(c => c.Id == id);
+       var userId = await GetUserId();
 
-       if (comment == null)
+        if (comment == null)
        {
-           throw new Exception("Comment not found");
+           throw new NullReferenceException(CommentNotFound);
        }
 
-       if (comment.AuthorId != dto.AuthorId)
+       if (comment.AuthorId != userId)
        {
-           throw new Exception("You are not the author of this comment");
+           throw new AccessViolationException(NotAuthorizedToEditComment);
        }
 
        comment.Content = dto.Content;
@@ -145,10 +155,16 @@ public class PublicationService : IPublicationService
     public async Task DeleteCommentAsync(Guid id)
     {
         var comment = await dbContext.Comments.FirstOrDefaultAsync(c => c.Id == id);
+        var userId = await GetUserId();
 
-        if(comment == null)
+        if (comment == null)
         {
-            throw new Exception("Comment not found");
+            throw new NullReferenceException(CommentNotFound);
+        }
+
+        if (comment.AuthorId != userId)
+        {
+            throw new AccessViolationException(NotAuthorizedToDeleteComment);
         }
 
         dbContext.Comments.Remove(comment);
@@ -159,19 +175,30 @@ public class PublicationService : IPublicationService
     {
         var publicationComments = await dbContext.Comments
             .Where(c => c.PublicationId == publicationId)
-            .OrderByDescending(c => c.DateCreated)
             .Select(c => new CommentDto()
             {
                 Id = c.Id,
                 Content = c.Content,
-                DateCreated = c.DateCreated.ToShortDateString() 
-                              + " " 
-                              + c.DateCreated.ToShortTimeString(),
+                DateCreated = c.DateCreated,
                 AuthorId = c.AuthorId,
                 PublicationId = c.PublicationId
             })
+            .OrderByDescending(c => c.DateCreated)
             .ToListAsync();
 
         return publicationComments;
+    }
+
+    private async Task<Guid> GetUserId()
+    {
+        var userId = httpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value!;
+        //var userId = await dbContext.Users.FirstOrDefaultAsync(u => u.Email == userEmail);
+
+        if (userId == null)
+        {
+            throw new NullReferenceException(PublicationAuthorNotFound);
+        }
+
+        return Guid.Parse(userId);
     }
 }
